First styles replaced with styled-components !! :D
