+++ components/BoxView/BoxView.style.js
@@ -1,8 +1,11 @@
-.boxView {
+import styled from "styled-components"
+import BoxView from "./BoxView.js"
+
+export const StyledBoxView = styled(BoxView)`
     display: grid;
     justify-items: center;
     align-items: center;
     overflow: hidden;
     width: 100%;
     height: 100%;
-}
+`
